Update command semantics

+++ src/main/java/seedu/binbash/command/DeleteCommand.java
@@ -3,6 +3,8 @@ package seedu.binbash.command;
 import java.util.regex.Pattern;
 import java.util.logging.Level;
 import seedu.binbash.ItemList;
+import seedu.binbash.ui.Ui;
+import seedu.binbash.storage.Storage;
 
 public class DeleteCommand extends Command {
     public static final Pattern COMMAND_FORMAT = Pattern.compile("delete\\s(?<identifier>.+)");
@@ -10,8 +12,7 @@ public class DeleteCommand extends Command {
     private int index;
     private boolean isIndex;
 
-    public DeleteCommand(ItemList itemList, int index) {
-        super(itemList);
+    public DeleteCommand(int index) {
         this.index = index;
         isIndex = true;
         commandLogger.fine(String.format(
@@ -20,8 +21,7 @@ public class DeleteCommand extends Command {
         ));
     }
 
-    public DeleteCommand(ItemList itemList, String keyword) {
-        super(itemList);
+    public DeleteCommand(String keyword) {
         this.keyword = keyword;
         isIndex = false;
         commandLogger.fine(String.format(
@@ -30,9 +30,12 @@ public class DeleteCommand extends Command {
         ));
     }
 
-    public boolean execute() {
+    public boolean execute(Ui ui, ItemList itemList, Storage storage) {
         if (isIndex) {
-            // Ensure index out of bounds error is caught by Parser.
+            if (index <= 0 || index > itemList.getItemCount()) {
+                executionUiOutput = "Index is out of bounds!";
+                return true;
+            }
             assert index > 0 && index <= itemList.getItemCount();
             commandLogger.log(Level.INFO, "Delete identifier is detected as an index");
             executionUiOutput = itemList.deleteItem(index);
@@ -40,6 +43,7 @@ public class DeleteCommand extends Command {
             commandLogger.log(Level.INFO, "Delete identifier is detected as an item name");
             executionUiOutput = itemList.deleteItem(keyword);
         }
+        storage.saveToStorage(itemList.getItemList());
         return true;
     }
 }
